acquisition: move to new namespace
